added owner-checking, edited search()-method

+++ src/main/java/com/sgkhmjaes/jdias/service/impl/ContactServiceImpl.java
@@ -1,14 +1,17 @@
 package com.sgkhmjaes.jdias.service.impl;
 
+import com.sgkhmjaes.jdias.domain.Person;
 import com.sgkhmjaes.jdias.service.ContactService;
 import com.sgkhmjaes.jdias.domain.Contact;
 import com.sgkhmjaes.jdias.repository.ContactRepository;
 import com.sgkhmjaes.jdias.repository.search.ContactSearchRepository;
+import com.sgkhmjaes.jdias.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
@@ -25,12 +28,16 @@ public class ContactServiceImpl implements ContactService{
     private final Logger log = LoggerFactory.getLogger(ContactServiceImpl.class);
 
     private final ContactRepository contactRepository;
+    private final UserService userService;
 
     private final ContactSearchRepository contactSearchRepository;
 
-    public ContactServiceImpl(ContactRepository contactRepository, ContactSearchRepository contactSearchRepository) {
+    public ContactServiceImpl(ContactRepository contactRepository,
+                              ContactSearchRepository contactSearchRepository,
+                              UserService userService) {
         this.contactRepository = contactRepository;
         this.contactSearchRepository = contactSearchRepository;
+        this.userService = userService;
     }
 
     /**
@@ -54,10 +61,22 @@ public class ContactServiceImpl implements ContactService{
      */
     @Override
     @Transactional(readOnly = true)
+    public List<Contact> findAllByUser() {
+        log.debug("Request to get all Contacts by user");
+        return new ArrayList<>(userService.getCurrentPerson().getContacts());
+    }
+
+  /*  /**
+     *  Get all the contacts.
+     *
+     *  @return the list of entities
+     *
+    @Override
+    @Transactional(readOnly = true)
     public List<Contact> findAll() {
         log.debug("Request to get all Contacts");
         return contactRepository.findAll();
-    }
+    }*/
 
     /**
      *  Get one contact by id.
@@ -69,7 +88,12 @@ public class ContactServiceImpl implements ContactService{
     @Transactional(readOnly = true)
     public Contact findOne(Long id) {
         log.debug("Request to get Contact : {}", id);
-        return contactRepository.findOne(id);
+
+        Contact contact = contactRepository.findOne(id);
+
+        if(userService.getCurrentPerson().getContacts().contains(contact))
+            return contact;
+        else return null;
     }
 
     /**
@@ -80,8 +104,13 @@ public class ContactServiceImpl implements ContactService{
     @Override
     public void delete(Long id) {
         log.debug("Request to delete Contact : {}", id);
-        contactRepository.delete(id);
-        contactSearchRepository.delete(id);
+
+        Contact contact = contactRepository.findOne(id);
+
+        if(userService.getCurrentPerson().getContacts().contains(contact)) {
+            contactRepository.delete(id);
+            contactSearchRepository.delete(id);
+        }
     }
 
     /**
@@ -94,8 +123,18 @@ public class ContactServiceImpl implements ContactService{
     @Transactional(readOnly = true)
     public List<Contact> search(String query) {
         log.debug("Request to search Contacts for query {}", query);
-        return StreamSupport
+
+        List<Contact> foundContactList = StreamSupport
             .stream(contactSearchRepository.search(queryStringQuery(query)).spliterator(), false)
             .collect(Collectors.toList());
+
+        List<Contact> usersContactList = new ArrayList<>(userService.getCurrentPerson().getContacts());
+
+        foundContactList.forEach(aspect -> {
+            if(!usersContactList.contains(aspect))
+                foundContactList.remove(aspect);
+        });
+        return foundContactList;
+
     }
 }
